+++ app/comments/router.py
@@ -2,19 +2,22 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette import status
-
 from app import User
+from app.adv.models import Group
 from app.adv.repository import AdvRepository
 from app.adv.service import get_adv_repo
 from app.auth.service import check_user_auth
 from app.comments.repository import CommentRepository
 from app.comments.schemas import CommentBase, CommentCreate
 from app.database import get_session
+from app.users.models import Roles
+from app.users.permissions import check_owner_or_admin, PermissionChecker
 
 comment_router = APIRouter()
 
 
 async def get_comment_repo(session: AsyncSession = Depends(get_session)) -> CommentRepository:
+    """Service function to return class with Comment CRUD operations"""
     return CommentRepository(session)
 
 
@@ -52,7 +55,10 @@ async def update_comments(adv_id: int,
                           comment_id: int,
                           comment_data: CommentCreate,
                           repo: CommentRepository = Depends(get_comment_repo),
-                          adv_repo: AdvRepository = Depends(get_adv_repo)):
+                          adv_repo: AdvRepository = Depends(get_adv_repo),
+                          current_user: User = Depends(check_user_auth)):
+    """Router for changing a comment. Accessible only for
+    AUTHOR of the comment and for ADMIN"""
     adv = await adv_repo.read(adv_id)
     if not adv:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
@@ -61,6 +67,7 @@ async def update_comments(adv_id: int,
     if not comment:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                             detail="Comment is not found")
+    check_owner_or_admin(comment, current_user)
     await repo.update(comment_data, comment)
     comments = await repo.read_all(adv_id)
     return comments
@@ -70,7 +77,10 @@ async def update_comments(adv_id: int,
 async def delete_comment(adv_id: int,
                          comment_id: int,
                          repo: CommentRepository = Depends(get_comment_repo),
-                         adv_repo: AdvRepository = Depends(get_adv_repo)):
+                         adv_repo: AdvRepository = Depends(get_adv_repo),
+                         current_user: User = Depends(check_user_auth)):
+    """Router for deleting a comment. Accessible only for
+     AUTHOR of the comment and for ADMIN"""
     adv = await adv_repo.read(adv_id)
     if not adv:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
@@ -79,5 +89,18 @@ async def delete_comment(adv_id: int,
     if not comment:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                             detail="Comment is not found")
+    check_owner_or_admin(adv, current_user)
     await repo.delete(comment_id)
     return f"Comment '{comment.body}' has been deleted successfully"
+
+
+@comment_router.delete('/del_comments/{adv_group}',
+                       dependencies=[Depends(PermissionChecker([Roles.ADMIN_ROLE]))])
+async def delete_comments(adv_group: Group,
+                         repo: CommentRepository = Depends(get_comment_repo)):
+    """Router for deleting all comments from specific Adv Group.
+    Accessible only for ADMIN"""
+    res = await repo.delete_group(adv_group)
+    if not res:
+        return f"No comments to delete from '{adv_group}' group advertisements"
+    return f"{res} comment(-s) deleted from '{adv_group}' group advertisements"
